feat: switch to bfloat16 and add generation sanity-check script
- Change model loading dtype from float32 -> bfloat16 in both
  multihost_extract.py (torch_dtype) and QwenConfig construction
- Change QwenConfig default dtype from jnp.float32 -> jnp.bfloat16
  so all JAX Dense layers and norms use bfloat16 by default
- Add test_generation.py: standalone greedy generation on a few English
  prompts to verify the JAX Qwen2 implementation end-to-end

This halves activation storage requirements and speeds up TPU compute.

+++ qwen2_jax.py
@@ -45,7 +45,7 @@ class QwenConfig:
     use_sliding_window: bool = False
     sliding_window: int = 32768
     use_fixed_cache: bool = False  # Use MaxText-style fixed-size cache
-    dtype: Any = struct.field(pytree_node=False, default=jnp.float32)
+    dtype: Any = struct.field(pytree_node=False, default=jnp.bfloat16)
 
 
 def rotate_half(x):
